requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
torch>=2.0.0
transformers>=4.35.0
TTS>=0.22.0
soundfile>=0.12.1
numpy>=1.22.0,<1.25.0
aiofiles>=23.2.1
pydantic>=2.4.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
watchdog>=3.0.0
 